fixing bugs

+++ app/controllers/likes_controller.rb
@@ -3,6 +3,7 @@ class LikesController < ApplicationController
   def create
     @bookmark = Bookmark.find(params[:bookmark_id])
     like = current_user.likes.build(bookmark: @bookmark)
+    authorize like
 
     if like.save
       flash[:notice] = "Bookmark Liked"
@@ -14,7 +15,8 @@ class LikesController < ApplicationController
 
   def destroy
     @bookmark = Bookmark.find(params[:bookmark_id])
-    like = current_user.likes.find(params[:id])
+    like = Like.find(params[:id])
+    authorize like
 
     if like.destroy
       flash[:notice] = "Unliked!"
